+++ lib/src/router_utils.dart
@@ -2,7 +2,7 @@ import 'package:flutter/widgets.dart';
 
 class RouterUtils {
   RouterUtils._();
-  static String getLocation(BuildContext context) {
+  static String getUrl(BuildContext context) {
     final r = Router.of(context);
     final parser = r.routeInformationParser;
     final config = r.routerDelegate.currentConfiguration;