+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+* Change from PathBuilder to Routed
+* Switch to RoutedWidgetSwitcher to RoutedSwitcher
+* Add support for inheritance and relative paths
+* Remove builder param
+
 ## 1.0.4
 * Added VRouter, Routemaster and Beamer implementations
 - Added tests for third party routers